+++ frontend/pong/js/components/match/PlayerProfile.js
@@ -1,12 +1,9 @@
 import { getParticipations } from "../../api/participations/getParticipations";
-import { BootstrapBadge } from "../../bootstrap/components/badge";
 import { BootstrapSpacing } from "../../bootstrap/utilities/spacing";
-import { BootstrapText } from "../../bootstrap/utilities/text";
 import { Component } from "../../core/Component";
-import { MatchEnums } from "../../enums/MatchEnums";
-import { createElement } from "../../utils/elements/createElement";
 import { createStartFlexBox } from "../../utils/elements/div/createFlexBox";
-import { ParticipationProfile } from "./ParticipationProfile";
+import { ParticipationProfile } from "../tournament/ParticipationProfile";
+import { createScore } from "./createScore";
 
 export class PlayerProfile extends Component {
   static #DEFAULT_HEIGHT = "max(15px, 3vh)";
@@ -45,25 +42,3 @@ export class PlayerProfile extends Component {
     this.append(container);
   }
 }
-
-const createScore = (scoreNum, matchResult) => {
-  const score = createElement("div", { textContent: `${scoreNum}` });
-
-  BootstrapText.setFontSize(score, 4);
-
-  switch (matchResult) {
-    case MatchEnums.Result.WIN:
-      BootstrapBadge.setSuccess(score);
-      break;
-    case MatchEnums.Result.LOSE:
-      BootstrapBadge.setDanger(score);
-      break;
-    case MatchEnums.Result.PENDING:
-      BootstrapBadge.setSecondary(score);
-      break;
-    default:
-      BootstrapBadge.setSecondary(score);
-      break;
-  }
-  return score;
-};
